# web/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "paths" : {
...
    "/api/chapters/{code}" : {
      "get" : {
        "tags" : [ "chapter-controller" ],
        "operationId" : "chapter",
        "parameters" : [ {
          "name" : "code",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ChapterOverview"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/chapters/{code}/sync" : {
      "get" : {
        "tags" : [ "chapter-controller" ],
        "operationId" : "sync",
        "parameters" : [ {
          "name" : "code",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ChapterSync"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/chapters/{code}/members" : {
      "get" : {
        "tags" : [ "chapter-controller" ],
        "operationId" : "members",
        "parameters" : [ {
          "name" : "code",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/MemberRow"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "PlannedChange" : {
        "type" : "object",
        "properties" : {
          "kind" : {
            "type" : "string"
          },
          "tool" : {
            "type" : "string"
          },
          "resource" : {
            "type" : "string"
          },
          "person" : {
            "type" : [ "string", "null" ]
          },
          "accountId" : {
            "type" : [ "string", "null" ]
          },
          "fromAccess" : {
            "type" : [ "string", "null" ]
          },
          "toAccess" : {
            "type" : [ "string", "null" ]
          }
        }
      },
...
      "ChapterOverview" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "code" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string"
          },
          "stats" : {
            "$ref" : "#/components/schemas/ChapterStats"
          },
          "registrationLink" : {
            "type" : "string"
          },
          "notionProblems" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "ChapterStats" : {
        "type" : "object",
        "properties" : {
          "activeMembers" : {
            "type" : "integer",
            "format" : "int32"
          },
          "alumni" : {
            "type" : "integer",
            "format" : "int32"
          },
          "liveProjects" : {
            "type" : "integer",
            "format" : "int32"
          },
          "leads" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "ChapterSync" : {
        "type" : "object",
        "properties" : {
          "tools" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ToolSync"
            }
          },
          "changes" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PlannedChange"
            }
          }
        }
      },
      "ToolSync" : {
        "type" : "object",
        "properties" : {
          "tool" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "mode" : {
            "type" : "string"
          },
          "ranAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "error" : {
            "type" : [ "string", "null" ]
          },
          "adds" : {
            "type" : "integer",
            "format" : "int32"
          },
          "changes" : {
            "type" : "integer",
            "format" : "int32"
          },
          "removals" : {
            "type" : "integer",
            "format" : "int32"
          },
          "drift" : {
            "type" : "integer",
            "format" : "int32"
          },
          "unmatchedAccounts" : {
            "type" : "integer",
            "format" : "int32"
          },
          "missingResources" : {
            "type" : "integer",
            "format" : "int32"
          },
          "creates" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "AccountSummary" : {
        "type" : "object",
        "properties" : {
          "tool" : {
            "type" : "string"
          },
          "state" : {
            "type" : "string"
          }
        }
      },
      "MemberRow" : {
        "type" : "object",
        "properties" : {
          "personId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "email" : {
            "type" : [ "string", "null" ]
          },
          "status" : {
            "type" : "string"
          },
          "kind" : {
            "type" : "string"
          },
          "chapterRole" : {
            "type" : [ "string", "null" ]
          },
          "title" : {
            "type" : [ "string", "null" ]
          },
          "projects" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "accounts" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AccountSummary"
            }
          },
          "claimed" : {
            "type" : "boolean"
          }
        }
      },
...
    }
  }
}